packages/sh_dendrite/pyproject.toml
[project]
name = "sh_dendrite"
version = "0.1.0"
description = "A super lightweight set of abstractions for building event-sourced applications"
readme = "README.md"
authors = [
    { name = "Howard Dierking", email = "howard@dierking.family" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.5",
    "opentelemetry-api>=1.38.0",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
